Änderungen aus xmi2db_norbit übernommen

+++ converter/compareschema.sh
@@ -1,47 +1,33 @@
 #!/bin/bash
 
 set -e
+[ -z "$UMLSCHEMA" ] && echo "UMLSCHEMA not found" && exit 1
+[ -f "/tmp/alkis-schema-$UMLSCHEMA.sql" ] || { echo "/tmp/alkis-schema-$UMLSCHEMA.sql not found"; exit 1; }
 
-pushd converter
-rm -f /tmp/xmi2db.log
+ORIGSCHEMA=aaa_orig
+NEWSCHEMA=aaa_new
 
-php-cgi -q db2ogr.php \
-	umlSchema=aaa_uml \
-	epsgCode=:alkis_epsg \
-	ogrSchema=':"alkis_schema"' \
-	withCodeLists=1 \
-	loglevel=2 \
-	>/tmp/alkis-schema.sql
-
-sed -e "s/<br>/\n/g" /tmp/xmi2db.log >/tmp/xmi2db-sql.log
-
-php-cgi -q db2gfs.php \
-	umlSchema=aaa_uml \
-	epsgCode=:alkis_epsg \
-	loglevel=2 \
-	>/tmp/alkis-schema.gfs
-
-sed -e "s/<br>/\n/g" /tmp/xmi2db.log >/tmp/xmi2db-gfs.log
-
-popd
-
-D=$PWD
-
-pushd ~/src/alkis-import
-psql -v alkis_epsg=25832 service=xmi2db <<EOF
-\set ON_ERROR_STOP
-DROP SCHEMA IF EXISTS aaa_orig CASCADE;
-CREATE SCHEMA aaa_orig;
-SET search_path = aaa_orig, public;
-\i $D/sql/alkis-schema-original.sql
-EOF
-popd
-
-psql -v alkis_epsg=25832 -v alkis_schema=aaa_xmi2db service=xmi2db <<EOF
+if ! [ -f .new-created ] || [ "/tmp/alkis-schema-$UMLSCHEMA.sql" -nt .new-created ]; then
+	if ! [ -f .orig-created ]; then
+		psql -q -v alkis_epsg=25832 -v parent_schema=public -v postgis_schema=public -v alkis_schema=$ORIGSCHEMA service=xmi2db <<EOF
 \set ON_ERROR_STOP
 DROP SCHEMA IF EXISTS :"alkis_schema" CASCADE;
 CREATE SCHEMA :"alkis_schema";
-\i /tmp/alkis-schema.sql
+SET search_path = :"alkis_schema", public;
+\i ~/src/alkis-import/alkis-schema.sql
+EOF
+		touch -r ~/src/alkis-import/alkis-schema.sql .orig-created
+	fi
+
+	psql -q -v alkis_epsg=25832 -v parent_schema=aaa_xmi2db -v postgis_schema=public -v alkis_schema=$NEWSCHEMA service=xmi2db <<EOF
+\set ON_ERROR_STOP
+DROP SCHEMA IF EXISTS :"alkis_schema" CASCADE;
+CREATE SCHEMA :"alkis_schema";
+SET search_path = :"alkis_schema", public;
+\i /tmp/alkis-schema-$UMLSCHEMA.sql
 EOF
 
-perl converter/compare.pl | tee /tmp/schema.diff
+	touch -r /tmp/alkis-schema-$UMLSCHEMA.sql .new-created
+fi
+
+perl converter/compare.pl $ORIGSCHEMA $NEWSCHEMA $UMLSCHEMA 2>&1 | tee /tmp/schema.diff
